requirements.txt
streamlit
selenium
watchdog
matplotlib
plotly
yfinance
beautifulsoup4
htmltext
prettify
regex
loguru
parsel
pydantic
fastapi
opencv-python
pillow
uvicorn
python-multipart